Composite defoliation series to determine outbreak events

outbreak() takes a defol object from defoliate_trees() and composites it into a site-level object. Function parameters allow the user to filter the tree-level series in various ways to optimize thresholds of what constitutes an "outbreak" level event recorded by the host trees.

outbreak(x, filter_perc = 25, filter_min_series = 3, filter_min_defol = 1)

Arguments

x

a defol object

filter_perc

the minimum percentage of defoliated trees to be considered an outbreak. Default is 25 percent.

filter_min_series

The minimum number of trees required for an outbreak event. Default is 3 trees.

filter_min_defol

The minimum number of trees recording a defoliation event. Default is 1 tree.

Value

A data.frame obr object for the site that includes all trees in the host defol object. Columns in the obr include:

year for every year in the set of host trees,

num_defol the number of trees recording a defoliation event,

percent_defol the percent of trees recording a defoliation,

num_max_defol the number of trees recording a maximum growth suppression (or peak of that event on that tree),

perc_max_defol the percent of trees at maximum defoliation,

mean_gsi the average of all trees growth suppression index (gsi),

mean_ngsi the average of all trees normalized growth suppression index (ngsi),

outbreak_status whether that year constitutes an outbreak based on the filters applied to the function.

Examples

data("dmj_defol")
head(outbreak(dmj_defol))
#>   year samp_depth num_defol perc_defol num_max_defol perc_max_defol mean_gsi
#> 1 1675          2         0        0.0             0              0   1.5303
#> 2 1676          2         0        0.0             0              0   1.2435
#> 3 1677          2         1       50.0             0              0   0.9222
#> 4 1678          2         1       50.0             0              0   0.9716
#> 5 1679          2         1       50.0             0              0   1.1373
#> 6 1680          3         1       33.3             0              0   0.7244
#>   mean_ngsi outbreak_status
#> 1    1.5145         not_obr
#> 2    0.7143         not_obr
#> 3   -0.1820         not_obr
#> 4   -0.0442         not_obr
#> 5    0.4198         not_obr
#> 6   -0.7303        outbreak