Outbreak statistics

Summary statistics for inferred outbreaks

Usage

outbreak_stats(x)

Arguments

x

An obr object after running outbreak()

Value

A data frame with descriptive statistics for each outbreak event determined by outbreak(), including:

"start" -- first year of outbreak

"end" -- last year of outbreak

"duration" -- length of outbreak (in years)

"n_df_start" -- number of trees defoliated at the start

"perc_df_start" -- percent of trees defoliated at the start

"max_df_obr" -- maximum number of trees in the outbreak during a single year

"yr_max_df" -- year with the maximum number of trees defoliated

"yr_min_ngsi" -- year with the lowest mean normalized growth suppression index (NGSI)

"min_gsi" -- minimum growth suppression index

"min_ngsi" -- minimum normalized gsi

Note

Certain statistics will be set to NA for the final outbreak event if there was an ongoing defoliation event (in which series_end_event = TRUE in defoliate_trees()). This is because the end of the outbreak remains unknown, so statistics such as duration cannot be calculated. Statistics pertaining to the start of the event are provided.

Examples

data("dmj_obr")
outbreak_stats(dmj_obr)
#>   start  end duration n_df_start perc_df_start max_df_obr yr_max_df yr_min_ngsi
#> 1  1680 1699       20          1          33.3          3      1690        1692
#> 2  1753 1769       17          6          46.2          7      1754        1755
#> 3  1825 1840       16         11          78.6         12      1831        1826
#> 4  1849 1865       17          7          46.7         13      1852        1853
#> 5  1881 1895       15          8          53.3         14      1886        1885
#> 6  1959 1970       12          7          41.2         15      1960        1965
#> 7  1987   NA       NA          9          52.9         15        NA          NA
#>   min_gsi min_ngsi
#> 1   0.121   -2.407
#> 2   0.343   -1.611
#> 3   0.500   -1.304
#> 4   0.252   -1.994
#> 5   0.262   -1.945
#> 6   0.328   -1.830
#> 7   0.378   -1.640